requirements.txt
pyyaml==6.0
pydantic==2.5.0
numpy==1.24.0
python-dateutil==2.8.2